backend/requirements.txt
redis==5.2.0
aioredis==2.0.1
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
motor==3.3.2
pymongo==4.5.0
razorpay==1.4.1
httpx==0.27.0
gTTS==2.5.1
python-dotenv==1.0.1
python-multipart==0.0.9
firebase-admin==6.5.0
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.2.0
fpdf2==2.7.8
aiofiles==23.2.1
pandas>=2.0.0
numpy>=1.24.0
scikit-learn
matplotlib
seaborn
